third_party/WebKit/LayoutTests/fast/speech/scripted/start-exception-expected.txt

Test getting an exception when calling start() in the Speech JavaScript API

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS 'webkitSpeechRecognition' in self is true
PASS webkitSpeechRecognition == null is false

doubleStart():
window.r = new webkitSpeechRecognition()
PASS r.start() did not throw exception.
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onstart
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onaudiostart
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onsoundstart
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onspeechstart
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onnomatch
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onspeechend
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onsoundend
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onaudioend
PASS r.start() threw exception InvalidStateError: An attempt was made to use an object that is not, or is no longer, usable..
onend
PASS r.start() did not throw exception.
r.abort()
onend
PASS successfullyParsed is true

TEST COMPLETE
